packages/pyc/pyproject.toml
[project]
name = "hangar-pyc"
version = "0.1.0"
description = "pyCycle gas turbine cycle analysis MCP server for agentic analysis"
requires-python = ">=3.11"
dependencies = [
    "hangar-sdk[all]",
    "om-pycycle>=4.2",
    "openmdao>=3.35",
    "numpy>=1.21",
    "scipy>=1.7",
    "matplotlib",
]

[project.scripts]
pyc-server = "hangar.pyc.server:main"
pyc-cli    = "hangar.pyc.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hangar"]

[tool.uv.sources]
hangar-sdk = { workspace = true }
# For development: use local upstream clone
om-pycycle = { path = "../../upstream/pyCycle", editable = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "golden_physics: physics-invariant golden tests",
    "parity: parity tests vs direct pyCycle API",
]
